Correção nos valores de entrada
Foi corrigido o problema no código que impedia a entrada de números com
parte decimal.

+++ Trabalhos/src/view/ViewSaidas.java
@@ -21,7 +21,7 @@ public class ViewSaidas{
 	private JList<String> lva11;
 	private JList<String> lva12;
 	
-	public void mostrar(int valor, int d1, int d2) {
+	public void mostrar(double valor, int d1, int d2) {
 		f.setLayout(null);
 		f.setSize(500, 400);
 		f.setResizable(false);
